Split ChannelEvent into two parts
Split former ChannelEvent class into two parts. The
main part is now called PixieEvent. PixieEvent
contains all information pertaining to the pixie
channel which fired. The new ChannelEvent contains
a pointer to a PixieEvent.

These changes are made so that all versions of the
scan code will eventually use the same underlying
event structure.
