+++ Scan/include/Unpacker.hpp
@@ -24,7 +24,7 @@
 #define MAX_PIXIE_CHAN 15
 #endif
 
-class ChannelEvent;
+class PixieEvent;
 class ScanMain;
 
 class Unpacker{
@@ -43,8 +43,8 @@ class Unpacker{
 
 	ScanMain *scan_main; /// Pointer to the ScanMain object responsible for reading spill data.
 
-	std::deque<ChannelEvent*> eventList; /// The list of all events in the spill.
-	std::deque<ChannelEvent*> rawEvent; /// The list of all events in the event window.
+	std::deque<PixieEvent*> eventList; /// The list of all events in the spill.
+	std::deque<PixieEvent*> rawEvent; /// The list of all events in the event window.
 
 	std::string message_head; /// Prefix used for text output.
